check if /parameters is there

+++ trainerai.sh
@@ -4,6 +4,12 @@ sed -i 's/`//g' ./devel/include/pose_estimation/Persons.h
 sed -i 's/`//g' ./devel/include/pose_estimation/Person.h
 sed -i 's/`//g' ./devel/include/pose_estimation/Bodypart.h
 sed -i 's/`//g' ./devel/include/pose_estimation/Pixel.h
-python3 docker/scripts/SecureGoogleDrive.py
-bash script.sh
-rm script.sh
+
+if [[ ! -d './src/fastpose/parameters/' ]]
+then
+    echo "'./src/fastpose/parameters/' does not exist on your filesystem."
+    python3 docker/scripts/SecureGoogleDrive.py
+    bash script.sh
+    rm script.sh
+fi
+
